Refecotry reduce unuse code
