+++ go.mod
@@ -1,8 +1,9 @@
 module stackdriver-monitoring-simple-reporter
 
 require (
-	cloud.google.com/go v0.32.0
+	cloud.google.com/go v0.34.0
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0
+	github.com/golang/lint v0.0.0-20180702182130-06c8688daad7 // indirect
 	github.com/googleapis/gax-go v2.0.0+incompatible // indirect
 	github.com/jung-kurt/gofpdf v1.0.0
 	github.com/wcharczuk/go-chart v2.0.1+incompatible
@@ -13,6 +14,6 @@ require (
 	google.golang.org/api v0.0.0-20181102150758-04bb50b6b83d
 	google.golang.org/appengine v1.3.0
 	google.golang.org/genproto v0.0.0-20181101192439-c830210a61df // indirect
-	google.golang.org/grpc v1.16.0 // indirect
+	google.golang.org/grpc v1.18.0 // indirect
 	gopkg.in/yaml.v2 v2.2.1
 )
